+++ bittytax/bittytax.py
@@ -182,12 +182,12 @@ def do_tax(transaction_records, tax_rules, skip_integrity_check):
 
     tax = TaxCalculator(transaction_history.transactions, tax_rules)
     tax.pool_same_day()
-    tax.match(tax.DISPOSAL_SAME_DAY)
+    tax.match_sell(tax.DISPOSAL_SAME_DAY)
 
     if tax_rules == config.TAX_RULES_UK_INDIVIDUAL:
-        tax.match(tax.DISPOSAL_BED_AND_BREAKFAST)
+        tax.match_buyback(tax.DISPOSAL_BED_AND_BREAKFAST)
     elif tax_rules in config.TAX_RULES_UK_COMPANY:
-        tax.match(tax.DISPOSAL_TEN_DAY)
+        tax.match_sell(tax.DISPOSAL_TEN_DAY)
 
     tax.process_section104(skip_integrity_check)
     return tax, value_asset