+++ CHANGELOG.md
@@ -6,6 +6,13 @@ Important:-
 
 2. The `Charity-Sent` transaction type has been changed from a normal disposal (same as a `Gift-Sent`) to being a "*No Gain/No Loss*" disposal, the same as a `Gift-Spouse`. If you have used this transaction type previously we recommend you re-generate your tax reports as you may have overpaid capital gains tax.
 
+3. The ExchangeRatesAPI and RatesAPI data sources are no longer available. Please update your `bittytax.conf` file to use the new BittyTaxAPI as shown below, this file resides in your .bittytax folder within your home directory.
+
+```
+data_source_fiat:
+    ['BittyTaxAPI']
+```
+
 ### Fixed
 - Accounting tool: "xlrd.biffh.XLRDError: Excel xlsx file; not supported" Exception. ([#36](https://github.com/BittyTax/BittyTax/issues/36))
 - Coinbase parser: added support for Convert transactions. ([#46](https://github.com/BittyTax/BittyTax/issues/46))
@@ -17,6 +24,9 @@ Important:-
 - Price tool: -ds option returns "KeyError: 'price'" exception.
 - Conversion tool: strip whitespace from header.
 - Accounting tool: Charity-Sent should be a "No Gain/No Loss" disposal. ([#77](https://github.com/BittyTax/BittyTax/issues/77))
+- Accounting tool: The "ten day" rule for companies, should match the buy to sell, not sell to buy-back. ([#131](https://github.com/BittyTax/BittyTax/issues/131))
+- Kraken parser: Trading pair split broken for XTZ/GBP. ([#124](https://github.com/BittyTax/BittyTax/issues/124))
+- Binance parser: Removed "Unexpected Coin content" error. ([#132](https://github.com/BittyTax/BittyTax/issues/132)) 
 ### Added
 - Etherscan parser: added internal transactions export.
 - Binance parser: added cash deposit and withdrawal exports.
@@ -58,6 +68,14 @@ Important:-
 - HitBTC parser: added new trade history export.
 - Electrum parser: new data file format added.
 - Accounting tool: added dictionary to TransactionRow.
+- New data source "Frankfurter" added for fiat exchange rates.
+- Ledger Live parser: new data file format added.
+- New data source "BittyTaxAPI" added for fiat exchange rates.
+- Coinfloor parser: new "trades" data file format added.
+- Gravity parser: new data file format added.
+- Etherscan parser: new "Transactions" data file format added.
+- Conversion tool: added parser for BscScan explorer.
+- Conversion tool: specify the local currency of the "Value" headers.
 - Conversion tool: added parser for Exodus wallet.
 ### Changed
 - Conversion tool: UnknownAddressError exception changed to generic DataFilenameError.
@@ -96,6 +114,8 @@ Important:-
 - Accounting tool: use fixed value (when specified) for counter asset prices.
 - Accounting tool: don't store fixed value for transfers.
 - Accounting tool: refactored import_records.py to use dictionary.
+- Binance parser: added "POS savings interest" and "Savings Interest" operations. ([#137](https://github.com/BittyTax/BittyTax/issues/137))
+- Binance parser: added "Super BNB Mining" operation.
 ### Removed
 - Accounting tool: skip audit (-s or --skipaudit) option removed.
 - Accounting tool: updated transactions debug removed.